+++ src/Raven.CodeAnalysis/BoundTree/Lowering/Lowerer.Invocation.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 
@@ -12,6 +13,12 @@ internal sealed partial class Lowerer
     {
         var receiver = (BoundExpression?)VisitExpression(node.Receiver);
         var arguments = node.Arguments.Select(a => (BoundExpression)VisitExpression(a)!).ToArray();
+        var staticQualifiedExtensionCall =
+            node.Method.IsExtensionMethod &&
+            node.ExtensionReceiver is null &&
+            receiver is BoundTypeExpression &&
+            arguments.Length == node.Method.Parameters.Length &&
+            node.Method.Parameters.Length > 0;
 
         var receiverCameFromInvocation = node.ExtensionReceiver is not null &&
             ReferenceEquals(node.ExtensionReceiver, node.Receiver);
@@ -49,19 +56,87 @@ internal sealed partial class Lowerer
 
                 _loweringTrace.RecordExtensionInvocation(traceEntry);
             }
-            return new BoundInvocationExpression(
+            var lowered = new BoundInvocationExpression(
                 node.Method,
                 loweredArguments,
                 receiver: null,
                 requiresReceiverAddress: node.RequiresReceiverAddress);
+            return HoistInvocationSubexpressionsIfNeeded(lowered);
         }
 
-        return new BoundInvocationExpression(
+        if (staticQualifiedExtensionCall)
+        {
+            var lowered = new BoundInvocationExpression(
+                node.Method,
+                arguments,
+                receiver: null,
+                extensionReceiver: null,
+                requiresReceiverAddress: node.RequiresReceiverAddress);
+            return HoistInvocationSubexpressionsIfNeeded(lowered);
+        }
+
+        var invocation = new BoundInvocationExpression(
             node.Method,
             arguments,
             receiver,
             extensionReceiver,
             node.RequiresReceiverAddress);
+        return HoistInvocationSubexpressionsIfNeeded(invocation);
+    }
+
+    private BoundExpression HoistInvocationSubexpressionsIfNeeded(BoundInvocationExpression invocation)
+    {
+        var receiver = invocation.Receiver;
+        var arguments = invocation.Arguments.ToArray();
+        if (!NeedsSubexpressionHoisting(receiver, arguments))
+            return invocation;
+
+        var compilation = GetCompilation();
+        var statements = new List<BoundStatement>();
+
+        BoundExpression? hoistedReceiver = null;
+        if (receiver is not null)
+            hoistedReceiver = HoistToTemp(receiver, "callReceiver", statements, compilation);
+
+        var hoistedArguments = new BoundExpression[arguments.Length];
+        for (var i = 0; i < arguments.Length; i++)
+            hoistedArguments[i] = HoistToTemp(arguments[i], $"callArg{i}", statements, compilation);
+
+        statements.Add(new BoundExpressionStatement(
+            new BoundInvocationExpression(
+                invocation.Method,
+                hoistedArguments,
+                hoistedReceiver,
+                extensionReceiver: null,
+                invocation.RequiresReceiverAddress)));
+
+        return compilation.BoundNodeFactory.CreateBlockExpression(statements);
+    }
+
+    private static bool NeedsSubexpressionHoisting(BoundExpression? receiver, BoundExpression[] arguments)
+    {
+        if (receiver is BoundBlockExpression)
+            return true;
+
+        for (var i = 0; i < arguments.Length; i++)
+        {
+            if (arguments[i] is BoundBlockExpression)
+                return true;
+        }
+
+        return false;
+    }
+
+    private BoundExpression HoistToTemp(
+        BoundExpression expression,
+        string nameHint,
+        List<BoundStatement> statements,
+        Compilation compilation)
+    {
+        var expressionType = expression.Type ?? compilation.ErrorTypeSymbol;
+        var temp = CreateTempLocal(nameHint, expressionType, isMutable: false);
+        statements.Add(new BoundLocalDeclarationStatement(
+            ImmutableArray.Create(new BoundVariableDeclarator(temp, expression))));
+        return new BoundLocalAccess(temp);
     }
 }
-